Beginning of proper list implementation
The list of S-Expressions seems to be a unique type in the literature (at least as far as Python is concerned). It's not fully-implemented, but I think I'm on the right track for getting it how I want.

Also includes my new TODO list. Pretty nifty, if I do say so myself.